backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
websockets>=12.0
python-multipart>=0.0.6
groq>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
